src/requirements.txt
discord.py>=2.1.0
requests>=2.28.1
termcolor>=2.2.0
chat-exporter>=2.5.1